// src/receiver/receiver-input.dto.ts
export class ReceiverInputDTO {
    id?: string;
    objectType?: string;
    orgId?: string;
    isDeletable?: boolean;
    status?: number;
    name?: string;
    nameAdd?: string;
    country?: string;
    postalCode?: string;
    city?: string;
    street?: string;
    email?: string;
    phone?: string;
    fax?: string;
    webSite?: string;
}